capi: Factor out util module
Move a bunch of utility functionality into the newly introduced util
module. In so doing we make CI test everything in this module under
Miri.
